Cleaned up the code
I removed almost all of the areas where there were more than 4 indents. I couldn't figure out how to change the guess(letter) function so that it only goes to a max of 4 indents. The again() function got a massive clean up and has a max of 4 indents. Also fixed a bug where the ui wouldn't show up for the first guess